fix view

+++ db/migrate/20251005075000_remove_temple_id_from_pujas.rb
@@ -1,5 +1,7 @@
 class RemoveTempleIdFromPujas < ActiveRecord::Migration[7.2]
   def change
-    remove_column :pujas, :temple_id, :bigint
+    if column_exists?(:pujas, :temple_id)
+      remove_column :pujas, :temple_id, :bigint
+    end
   end
 end
